// frontend/src/app/api/model/mealDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DishDto } from './dishDto';


export interface MealDto { 
    id?: string;
    mealType: MealDto.MealTypeEnum;
    dateTime: string;
    dishes: Array<DishDto>;
}
export namespace MealDto {
    export const MealTypeEnum = {
        Breakfast: 'BREAKFAST',
        Lunch: 'LUNCH',
        Dinner: 'DINNER',
        Snack: 'SNACK'
    } as const;
    export type MealTypeEnum = typeof MealTypeEnum[keyof typeof MealTypeEnum];
}